requirements.txt
underthesea==9.2.11
mysql-connector-python==9.6.0
python-dotenv==1.2.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
pandas==2.2.3
numpy==1.26.4
requests==2.33.0